Stützstellen werden jetzt korrekt ausgeliefert, nicht mehr mit Faktor 1000 multipliziert. Neuer Anfragekonstruktor für nicht-zyklische Anfragen. Prognoseganglinie umfasst jetzt nur noch den angefragten Bereich.
git-svn-id: svn://ntoserver/Datenverteiler/VRZ3/trunk/de.bsvrz.iav.gllib@2394 40c0be50-5d29-0410-901d-fb59edb5cafc

+++ src/de/bsvrz/iav/gllib/gllib/GanglinienOperationen.java
@@ -401,7 +401,7 @@ public class GanglinienOperationen {
 		} else {
 			ende = g2.stuetzstellen.lastKey();
 		}
-		breite = Math.round((float) (ende - start) / intervalle);
+		breite = Math.round((double) (ende - start) / intervalle);
 
 		// Haltepunkte bestimmen
 		zeitstempel = new TreeSet<Long>();
